requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
greenlet==1.1.2
idna==3.3
numpy==1.22.1
oauthlib==3.1.1
pandas==1.4.0
pika==1.2.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.31
tweepy==4.5.0
urllib3==1.26.8
